+++ internal/migrate/migrations/0004_create_location_table.sql
@@ -1,10 +1,10 @@
 CREATE TYPE kind AS ENUM ('s3', 'ftp', 'ftps', 'sftp');
 
 CREATE TABLE location (
-    id serial PRIMARY KEY,
+    id bigserial PRIMARY KEY,
     kind kind NOT NULL,
     info jsonb NOT NULL,
-    project_id integer NOT NULL REFERENCES project(id) ON DELETE CASCADE,
+    project_id bigint NOT NULL REFERENCES project(id) ON DELETE CASCADE,
 
     -- metadata columns
     created_at timestamptz NOT NULL DEFAULT NOW(),
