+++ internal/migrate/migrations/0005_create_transfer_table.sql
@@ -1,9 +1,9 @@
 CREATE TABLE transfer (
-    id serial PRIMARY KEY,
+    id bigserial PRIMARY KEY,
     pattern text NOT NULL,
-    src_id integer NOT NULL REFERENCES location(id) ON DELETE CASCADE,
-    dst_id integer NOT NULL REFERENCES location(id) ON DELETE CASCADE,
-    project_id integer NOT NULL REFERENCES project(id) ON DELETE CASCADE,
+    src_id bigint NOT NULL REFERENCES location(id) ON DELETE CASCADE,
+    dst_id bigint NOT NULL REFERENCES location(id) ON DELETE CASCADE,
+    project_id bigint NOT NULL REFERENCES project(id) ON DELETE CASCADE,
 
     -- metadata columns
     created_at timestamptz NOT NULL DEFAULT NOW(),
