+++ internal/migrate/migrations/0006_create_schedule_table.sql
@@ -1,8 +1,8 @@
 CREATE TABLE schedule (
-    id serial PRIMARY KEY,
+    id bigserial PRIMARY KEY,
     name text NOT NULL,
     expr text NOT NULL,
-    project_id integer NOT NULL REFERENCES project(id) ON DELETE CASCADE,
+    project_id bigint NOT NULL REFERENCES project(id) ON DELETE CASCADE,
     UNIQUE (name, project_id),
 
     -- metadata columns
